revise to depend on transformers package
As discussed on the libraries list Sep-Nov 2010.  The mtl is split in two:

transformers is a Haskell 98 package containing
 * base functors (Data.Functor.Constant and Data.Functor.Identity),
 * operations on functors (Data.Functor.Compose and Data.Functor.Product),
 * transformer classes (Control.Monad.Trans.Class and Control.Monad.IO.Class)
   and
 * concrete monad transformers with code to lift operators
   (Control.Monad.Trans.*). 

The package can be used on its own (see the Control.Monad.Trans.Class
documentation for examples), or with packages adding type classes.

The revised mtl depends on transformers and adds type classes using
functional dependencies.  It has the same modules as version 1.*
and usage is very close, except for the differences listed below:

 * instances of Applicative and Alternative have been added as appropriate.
 * Functor instances for monad transformers no longer require Monad where
   Functor is sufficient.
 * simple monads are now aliases for monad trasformers applied to Identity.
 * The instance Error String is restructured to avoid a type synonym instance.